Implement switching between app & bootloader
This design assumes that the bootloader starts after AVR reset
(which is enabled by BOOTRST fuse). The bootloader decides if it
should execute itself or the application. The bootloader is only
executed if the reset was caused by a watchdog, and the magic key
is not set.

Add a new report to application that can be written to by the host
to cause the device to enter bootloader. The host can differentiate
whether the device runs application or bootloader by USB product ID.