Run app-avr with watchdog timer enabled
Since the only way to enter bootloader is to send a command to
the application, flashing a buggy or unresponsive application
might make it impossible to enter the bootloader, which can only
be fixed with a serial programmer.

Since watchdog reset is already used to enter bootloader, the app
can as well be always running under watchdog to enter bootloader
in case something goes wrong. For a false-positive watchdog reset,
reconnecting the device will load the app again.